Refactor: Move TaskRepository class into repository/
