+++ app/src/main/java/com/codesoom/assignment/repository/TaskRepository.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment;
+package com.codesoom.assignment.repository;
 
 import com.codesoom.assignment.models.Task;
