+++ app/src/main/java/com/codesoom/assignment/services/PostService.java
@@ -2,7 +2,7 @@ package com.codesoom.assignment.services;
 
 import com.codesoom.assignment.HttpStatusCode;
 import com.codesoom.assignment.JsonConverter;
-import com.codesoom.assignment.TaskRepository;
+import com.codesoom.assignment.repository.TaskRepository;
 import com.codesoom.assignment.models.Task;
 import com.sun.net.httpserver.HttpExchange;
